PR review updates: - merge charm lib files - rename interface
