+++ charms/istio-pilot/metadata.yaml
@@ -18,7 +18,7 @@ provides:
     schema: https://raw.githubusercontent.com/canonical/operator-schemas/master/ingress-auth.yaml
     versions: [v1]
   gateway:
-    interface: gateway
+    interface: istio-gateway-name
     description: |
       Provides gateway name related Juju application
 assumes:
